+++ tests/execveat.c
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2022 Gaël PORTAY
+ *
+ * SPDX-License-Identifier: LGPL-2.1-or-later
+ */
+
+#include <fcntl.h>
+#include <unistd.h>
+
+int main(void)
+{
+	char * const argv[] = { "-sh", "-c", "echo \"$@\"", "sh", "one", "two",
+				"three", NULL };
+
+	execveat(AT_FDCWD, "/bin/sh", argv, environ, AT_EMPTY_PATH);
+	_exit(127);
+}